fullDetails customer screen now functionally working with React Navigation.
