+++ src/components/CustomerFullDetails.js
@@ -2,6 +2,6 @@ import React from 'react';
 import { View, Text } from 'react-native';
 
 
-export const CustomerFullDetails = (id) => {
-   return <View><Text>This is the full details.{id}</Text></View>;
+export const CustomerFullDetails = (props) => {
+   return <Text>This is the full details for : {props.navigation.state.params.customer.name.first}</Text>;
 };
